+++ CHANGELOG.md
@@ -10,7 +10,10 @@ _None_
 
 ### New Features
 
-_None_
+* Files: Add new parser for accessing any files specified from the bundle
+  [Mike Gray](https://github.com/mgray88)
+  [#665](https://github.com/SwiftGen/SwiftGen/issues/665)
+  [#734](https://github.com/SwiftGen/SwiftGen/pull/734)
 
 ### Bug Fixes